requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tqdm>=4.60.0
seaborn>=0.11.0
timm>=0.5.4  # Para modelos pre-entrenados adicionales 